# Tidy up SpecialProperties utils