- Renamed `stripSpecialSpecialProperties` -> `stripAllSpecialProperties`
- Reformatted ifs for clarity
- Pre-allocate new body capacity based on maximum possible size (no stripped properties)
```
21:23 $ benchcmp -best old.txt new.txt
benchmark                                         old ns/op     new ns/op     delta
BenchmarkStripSpecialProperties/no_special-12     1657          903           -45.50%
BenchmarkStripSpecialProperties/special-12        1689          944           -44.11%

benchmark                                         old allocs     new allocs     delta
BenchmarkStripSpecialProperties/no_special-12     4              2              -50.00%
BenchmarkStripSpecialProperties/special-12        4              2              -50.00%

benchmark                                         old bytes     new bytes     delta
BenchmarkStripSpecialProperties/no_special-12     2143          1202          -43.91%
BenchmarkStripSpecialProperties/special-12        2143          1202          -43.91%
```